requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.0.2
openai==1.3.7
google-generativeai==0.3.2
requests==2.31.0
httpx>=0.24.0,<0.25.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
stripe==7.8.0
python-dotenv==1.0.0
apscheduler==3.10.4
redis==5.0.1
aioredis==2.0.1
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
